[expo-router] fix link preview navigation into nested stack
