+++ packages/expo-router/src/navigationParams.ts
@@ -1,3 +1,5 @@
+import { deepEqual } from './utils/deepEqual';
+
 export const INTERNAL_EXPO_ROUTER_NO_ANIMATION_PARAM_NAME = '__internal_expo_router_no_animation';
 export const INTERNAL_EXPO_ROUTER_IS_PREVIEW_NAVIGATION_PARAM_NAME =
   '__internal__expo_router_is_preview_navigation';
@@ -135,4 +137,15 @@ export function removeInternalExpoRouterParams(
     return undefined;
   }
   return removeParams(params, [...internalExpoRouterParamNames, 'params']);
+}
+
+export function areParamsEqualDisregardingInternalExpoRouterParams(
+  paramsA: Record<string, unknown> | object | undefined,
+  paramsB: Record<string, unknown> | object | undefined
+) {
+  return deepEqual(
+    // using ?? {}, because from our perspective undefined === {}, as both mean no params
+    removeInternalExpoRouterParams(paramsA ?? {}),
+    removeInternalExpoRouterParams(paramsB ?? {})
+  );
 }
