+++ packages/expo-router/src/utils/deepEqual.ts
@@ -1,0 +1,16 @@
+export function deepEqual(
+  a: { [key: string]: any } | undefined,
+  b: { [key: string]: any } | undefined
+): boolean {
+  if (a === b) {
+    return true;
+  }
+  if (a == null || b == null) {
+    return false;
+  }
+  if (typeof a !== 'object' || typeof b !== 'object') {
+    return false;
+  }
+  const keys = Object.keys(a);
+  return keys.length === Object.keys(b).length && keys.every((key) => deepEqual(a[key], b[key]));
+}
